Add files via upload
Working more on Grammar and Rule ideas. Sorting rules by 'abstractness' seems to be potentially useful for comparing languages and merging them. In addition can catch potentially undesired cyclic behavior, and advise config controls. Been grasping the issue the past few days, as I'd love to allow for EBNF within the cases and or REGEX for terminal strings. Both would complicate sorting cases, thus an initial algorithm to figure out the 'rank' or the min, max, [in terms of terminals/atoms] and depth of each rule name and case would allow a sortation of sorts of both rules and cases. The cases then can be evaluated with a lookup of a rule, and if a REGEX or EBNF form occurs within the case tokens, then the priority of that particular cases 'similarity' to that of other's in it's subsection can be adjusted. For example 'I love Candy' and 'I love [Legos|Candy]' and 'I love {toys}+' where toys is itself is rule object that ends with terminals, we may have a case where Legos and Candy result in similar terminals, and/or toys [which accepts 1 or more cases], in that case a warning should be raised when sorting stating 'hey man, they all have similar mins, but 'toys' has a potentially longer case, so that's going first, but ambiguity exists so check it out!'. This may be fixed with prefixing the rule Name along with a () literal strings, as a potential fix, similarly the sorting of the 'I love' will then result in Candy being compared to [Legos|Candy]. I'm conflicted on whether REGEX should be allowed, and if so whether redundant rules should be removed via keeping REGEX or removing them in favor of explicit cases. 
REGEX would also complicate other thoughts for the future of the project. 

TLDR: Trying to sort rules and ensure no cycles, and may help sort cases by a more substantial 'length' for it's pattern matching, and warn of ambiguity, though it's a problem that may be uglier than this to solve. The upside is the deeper investigations would only be alluded to before a full comparison. Problems regarding ambiguity might try being solved [within a singular grammar file or grammar collection] by prefixing with the ruleName, as the ruleName should be unique. Additionally to ensure the easiest use by a user, only ATOMS and terminals must be declared as rule names in some way, along with a meta data [if provided]. Thus the algorithm I'm working on is only working on that premise. Again ease of accessibility requires no rigor in crafting the grammar and only some idea of the 'flow' and should raise warnings for problems that go against theory.