Add files via upload
Detailed more on the problem to handle, and am working on a data structure to approach on idea. There may be algorithms out there people have researched, but I don't have access to their papers, or a budget. Hence I'll try to make some general structures really nicely with safety features to catch misuse and flexibility to control the encoding of each unit.

I'm going to try using a matrix with a legend to try getting around the EBNF form when just trying to find depths and min/max atoms/terminals of a given case path. This algorithm will likely explore the grammar densely, but as it should only run on a 'new' grammar/language, or after a grammar merge to find issues, it isn't the end of the world if it is expensive.