Add files via upload
Working on validating, thinking about how to make a few good example objects, and then loop through them.
Going to be tutoring a friend in Trig soon so stopping for now.

Accomplished some more logic validation on the member functions, and determined the way I want to have the object declared. 
The print is semi-done. It'll be enough for now at least. Ideally I'd like to have an empty matrix print a "", instead of "\n\n\n" 
Will likely finish validating the other functional features and then bother with the string.
I opted to go for a list of lists style matrix [instead of an array with the length Col*Row], as for column operations it'd be faster, and allocating additional chunks of memory shouldn't be *terrible*, a change to the Matrix structure will only change isPosition/getPosition/setPositions' implementation I think from the top of my head. That is an additional validation I'll have to do later tonight.

Strict Mode is intended to force very very strict declarations and uses of this data object, I've yet to decide if it's quite how I want it [pretty sure it isn't].