Add files via upload
Mixed up the validate column and row statements (literally just the last if statements in each needed to be swapped). Continuing to make this Matrix robust [do it once nicely]. I want to add an optional strict typing for column and row that goes with the legend(s). In this way the matrix can be controlled even tighter. The optional functions then can be forgotten about by the user and managed by the matrix object. 

When expanding the legend, I will make an executive decision that in strict mode, an optional NA requirement will be given, or if expanding the matrix with rows from other matrices. I'm still aways from this.

The relational DB designs also make me want to allow for populating from other matrices based on a key pairing, but that would require a parent object to manage the relations and privileges of each matrix. Again just wanted to write down some thoughts.