Add files via upload
Some more changes, ensuring logic goes well. Small bug once again fixed. Will test a bit more later tonight. After that some code organization, comment clean up, and implementing both strong typing and validation for when expanding the legend is desired. 
Validation for unique column names and row names also.

The validation may require unique values in each row/column. In this case I may have to research how functools may help. If the validate function cannot see the matrix container object, then there may need to be either a reference to the matrix that is immutable in a wrapper function, or functools may allow for the function's references to the matrix to be passed without a wrapper. 
Additionally validation with set row and with add row, along with when to expand the legend or default expand it is still up for debate. 
I think the legends should default expand, based on the index in the matrix, if the index doesn't already exist as a key, otherwise iterate until an available index is available as a legend key [to account for random legend naming and dim adding]. 

A rename in the legend is required, along with a revalidate/replace validation function.