Add files via upload
Simple bugs. Functools if not specified will assume arguments are in order of parameters, ending with two or more arguments potentially sent to the same parameter. 
Another bug was in the Remove functions - not fixed yet - simple fix though, just set the matrix. Keep thinking the changes will be maintained by the object. Will create a validate and set Matrix function. The validate simply ensures the matrix's column and rows are regular. 
I may be able to get away with a temporary container to reset the matrix without as much memory being used however.