Add files via upload
The actual bug with remove was just messing up a simple if statement, I was checking for index above 0 instead of above -1. Then as any columns added to the matrix are mutable beyond just the object [for example, passing in x=[a,b,c] into a matrix, then removing a row (so 'a' in the matrix is gone), it will affect the matrix, and vise versa. As such a copy.deepcopy is used. In strict mode I will try to ensure that copies are made and consumed, or that the original argument is consumed so it's more explicit that memory is being used.