Add files via upload
Deleted some comments from the main branch, just to focus more deeply on the legend feature. 
The legend feature I'm still wrestling with exactly how I want it to go down. I'm pretty sure I'll want to write a 'get default string' function, that finds an available string based on a formulation of how many entries in the idx, modulo 26, A*Divisor amt + Ascii Remainder. Similarly when removing an entry from the legend, I will want to keep the int idx's as low as possible. Finally, instead of a parent object, I'd like for an internal function to take the place [not hard to do, was just thinking about whether having a parent would be beneficial]. 
After those and some set features, and a strict mode, will integrate it with the matrix object and see where I need to add options to query the legend and get information that way.

Rough plan Legend -> integrate w/ Matrix -> Matrix used as a general purpose graph for  ruleObj rating depth and min/max atoms/terminals -> EBNF or Regex parsing of each case to somehow structure []+ and {}+ cases, or s.e cases, and somehow give additional 'points' to unknown length structures, and how early they appear in the string.  -> edit ruleObject to have cases keep their predicted rating (as rule objects are the results of their cases) -> Figure out whether there is a nice theory that holds true for ordering the cases by this logic [or find a good counter example] -> option A/B

A) Self Theory works, no research required, at least convincing enough to work, rewrite a file with the sorted cases. And a meta data sort file with the resulting stuff. -> create a seperate workflow for theory validation files, and a workflow for applying the grammar to a language [oop or functional, or a mix?] -> Pytorch tools for Langauge commonalities and maybe a NN to obtain phrases with some commonality. Would have to make a GUI interface to have the NN report it's predictive findings back to the user to see if it's reasonable. And if not, what caused the odd finding. It may be the case that certain works group together differently based on phrasing and time period written. In that case it'd be good to have separate models that one can edit to 'clean up' the inputs to be phrased how the 'main' model would like it.

Finally from there, one simply has to think how - in the card game case -how the cards should raise flags to a parent container, and what kind of history stacks or other data objects would need to occur. Card games mostly use Queues, Stacks, or a round robin for its features, so that might give a clue.

B) Self Theory doesn't work, research required. Or EBNF parser and rating gives issues, or there is too much cyclicsm to get much use out of the rating or depths. 


In total, it's a shot in the dark, but if I'm going to make it at all, I'd like to make it once a fully featured, so that even if the rule object algorithms don't turn out nicely, I've got a sweet couple of objects out of the deal.