Add files via upload
Sort Rule Legend is focusing on implementing Legend as a subobject, with as much as its own self validation as possible and the desired amount of freedom to be self contained. Still a work in progress, but not the main hold up.

EBNF skeleton is a naive approach at detecting EBNF forms. Ideally it will eventually be able to take care of each case along with |,+, *, ., and % symbols for additional functionalities, or detect Regex. The skeleton that validates that the cases are valid EBNF and that items not rules, literals, and atomics are raised as errors or misformated EBNF . 
From there, we can sort rules/cases using the Matrix and a simple algorithm [that should be something like Node*Edges for time complexity, and N*N 1/2 for space complexity,)

The EBNF skeleton will also [when it's finished] inform when matching cases, what to expect for each atom, literal, or rule name by detecting those things first, and being able to navigate cases with a lookahead. Finally a simple memoize function will then complete the first hurdle of the project.

From there some fun stuff really begins. More theory but the tools should be flexible enough to handle it [ignoring necessary meta data]. 

As I have a potential lead to employment, I'll begin shifting my focus back to Java, though on occasion I'll continue working on the EBNF.


The EBNF Skeleton has two errors in it's basic tests for its basic detection, shouldn't take too long to fix,, but I've got tutoring to do tonight.